requirements.txt
Pillow==9.0.1
psutil==5.9.0
qrcode==7.3.1
aiohttp==3.8.1
fastapi==0.75.0
uvicorn==0.17.5
apistats==0.4.1
asyncpraw==7.5.0
colorthief==0.2.1
fast_api_stats==0.0.2
python-dotenv==0.19.2
python-multipart==0.0.5